backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
motor>=3.3.0
python-dotenv>=1.0.0
pydantic>=1.10.7
python-multipart>=0.0.6
gunicorn>=20.1.0
geojson>=3.0.1
pytest>=7.3.1
httpx>=0.24.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
rtree>=1.0.1
libpysal>=4.10.0
esda>=2.5.0
mapclassify>=2.6.0
numpy>=1.24.0
pandas>=2.0.0
statsmodels>=0.14.0
fiona>=1.9.0
bcrypt<4.0.0
